+++ src/main/java/fr/avenirsesr/portfolio/backoffice/BackOfficeApplication.java
@@ -5,10 +5,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
-@SpringBootApplication(scanBasePackages = {
-    "fr.avenirsesr.portfolio.backoffice",
-    "fr.avenirsesr.portfolio.common"
-})
+@SpringBootApplication(
+    scanBasePackages = {"fr.avenirsesr.portfolio.backoffice", "fr.avenirsesr.portfolio.common"})
 public class BackOfficeApplication {
 
   public static void main(String[] args) {