+++ src/main/java/fr/avenirsesr/portfolio/backoffice/shared/domain/model/EConfiguration.java
@@ -1,5 +1,0 @@
-package fr.avenirsesr.portfolio.backoffice.shared.domain.model;
-
-public interface EConfiguration {
-  String name();
-}