+++ src/main/java/fr/avenirsesr/portfolio/backoffice/trace/domain/model/TraceConfiguration.java
@@ -1,4 +1,0 @@
-package fr.avenirsesr.portfolio.backoffice.trace.domain.model;
-
-public record TraceConfiguration(
-    int maxRemainingDays, int maxRemainingDaysBeforeWarning, int maxRemainingDaysBeforeCritical) {}