+++ src/main/java/fr/avenirsesr/portfolio/backoffice/trace/domain/port/input/TraceConfigurationService.java
@@ -1,6 +1,6 @@
 package fr.avenirsesr.portfolio.backoffice.trace.domain.port.input;
 
-import fr.avenirsesr.portfolio.backoffice.trace.domain.model.TraceConfiguration;
+import fr.avenirsesr.portfolio.common.configuration.domain.model.TraceConfiguration;
 
 public interface TraceConfigurationService {
   TraceConfiguration getTraceConfiguration();