+++ src/main/java/fr/avenirsesr/portfolio/backoffice/websitecontent/domain/model/EWebsiteContentConfiguration.java
@@ -1,6 +1,6 @@
 package fr.avenirsesr.portfolio.backoffice.websitecontent.domain.model;
 
-import fr.avenirsesr.portfolio.backoffice.shared.domain.model.EConfiguration;
+import fr.avenirsesr.portfolio.common.configuration.domain.model.EConfiguration;
 
 public enum EWebsiteContentConfiguration implements EConfiguration {
   BUILD_LIFE_PROJECT_CONTENT,