+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and [Conventional Commits](https://www.conventionalcommits.org/) standard.
 
 ---
 
+## [v1.0.1] - 2025-10-01
+
+- Fix:
+    - spring security configuration.
+    - some common objects moved to avenirs-portfolio-common repository.
+
 ## [v1.0.0] - 2025-09-23
 
 - Initial version, the features were moved from avenirs-portfolio-api and reorganized.